DB保存するかの判定導入

+++ backend/src/services/eventBus.ts
@@ -67,7 +67,12 @@ export class EventBus {
     });
   }
 
-  public async log(platform: Platform, color: Color, content: string) {
+  public async log(
+    platform: Platform,
+    color: Color,
+    content: string,
+    isSave: boolean = false
+  ) {
     const logEntry: ILog = {
       timestamp: new Date(),
       platform,
@@ -75,26 +80,28 @@ export class EventBus {
       content,
     };
 
-    try {
-      await Log.create(logEntry);
-      // 10000件を超える場合に5000件を削除
-      const logCount = await Log.countDocuments();
-      if (logCount > 10000) {
-        const logsToDelete = logCount - 5000;
-        const oldestLogs = await Log.find()
-          .sort({ timestamp: 1 })
-          .limit(logsToDelete);
+    if (isSave) {
+      try {
+        await Log.create(logEntry);
+        // 10000件を超える場合に5000件を削除
+        const logCount = await Log.countDocuments();
+        if (logCount > 10000) {
+          const logsToDelete = logCount - 5000;
+          const oldestLogs = await Log.find()
+            .sort({ timestamp: 1 })
+            .limit(logsToDelete);
 
-        if (oldestLogs.length > 0) {
-          await Log.deleteMany({
-            _id: { $in: oldestLogs.map((log) => log._id) },
-          });
-          console.log(`${logsToDelete}件の古いログを削除しました`);
+          if (oldestLogs.length > 0) {
+            await Log.deleteMany({
+              _id: { $in: oldestLogs.map((log) => log._id) },
+            });
+            console.log(`${logsToDelete}件の古いログを削除しました`);
+          }
         }
+        console.log('Log saved to database');
+      } catch (error) {
+        console.error('Error saving log:', error);
       }
-      console.log('Log saved to database');
-    } catch (error) {
-      console.error('Error saving log:', error);
     }
 
     this.publish({
